backend/requirements.txt
django>=5.0
djangorestframework>=3.15
django-cors-headers>=4.3
djangorestframework-simplejwt>=5.3
drf-spectacular>=0.27
python-decouple>=3.8
google-genai>=1.0
reportlab>=4.2
Pillow>=10.4
